chore(refactor): Cleans up build
