+++ app/assets/icons/DecorationIcon.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-const ButtonDecoration: FC = () => {
+const DecorationIcon: FC = () => {
   return (
     <svg
       xmlns='http://www.w3.org/2000/svg'
@@ -98,4 +98,4 @@ const ButtonDecoration: FC = () => {
   );
 };
 
-export default ButtonDecoration;
+export default DecorationIcon;
